Moved reaction button code into reaction-game.tsx
